refactor(library/URLComponent): moves out of "/customTypes" directory
