+++ src/features/TextToImage/config/utilities/DynamicConfig/FetchingError.ts
@@ -2,7 +2,7 @@ import Attempt from '@/library/Attempt';
 
 import type {
   URLPath,
-} from '@/library/customTypes/URLComponent';
+} from '@/library/URLComponent';
 
 class DynamicConfig_FetchingError
   extends Attempt.ActionableError<
