+++ src/features/TextToImage/webAPI/Server/ServerSpecificationError.ts
@@ -2,7 +2,7 @@ import Attempt from '@/library/Attempt';
 
 import type {
   URLPath,
-} from '@/library/customTypes/URLComponent';
+} from '@/library/URLComponent';
 
 class TextToImage_Server_SpecificationError
   extends Attempt.ActionableError<
